Tests/LibWeb/Text/input/inspector-overlay-visual-context-tree-growth.html

<!DOCTYPE html>
<!-- Inspector overlays used to append transient copies of the highlighted element's transform/scroll contexts onto
     the accumulated visual context tree on every recording, which had to be pruned again by the next recording to
     avoid unbounded growth (see #10368). Overlay commands now reference the element's existing context and apply it
     in geometry-only mode, so recording an overlay must not modify the tree at all.

     NB: The tree is only rebuilt from scratch when layout is invalidated. Mutating the DOM between measurements (e.g.,
     via println) would dirty layout and force a rebuild — masking accidental growth. So, as in
     visual-context-value-only-updates.html, all measurements are taken first, and printed at the end. -->
<style>
    #box {
        width: 100px;
        height: 100px;
        transform: rotate(10deg);
        background: orange;
    }
</style>
<script src="include.js"></script>
<div id="box"></div>
<script>
    test(() => {
        const box = document.getElementById("box");

        // dumpDisplayList() records a fresh display list; the count that follows is the resulting live visual context
        // tree size.
        const recordAndCount = () => {
            internals.dumpDisplayList();
            return internals.visualContextTreeNodeCount();
        };

        const baseCount = recordAndCount();

        internals.setHighlightedNode(box);
        const highlightedCounts = [];
        for (let i = 0; i < 6; i++)
            highlightedCounts.push(recordAndCount());

        internals.setHighlightedNode(null);
        const clearedCount = recordAndCount();

        const untouchedWhileHighlighted = highlightedCounts.every(c => c === baseCount);
        const untouchedAfterClearing = clearedCount === baseCount;

        println(untouchedWhileHighlighted
            ? "PASS: recording with an inspector highlight does not modify the visual context tree"
            : `FAIL: visual context tree changed while highlighted: base ${baseCount}, got ${highlightedCounts.join(", ")}`);
        println(untouchedAfterClearing
            ? "PASS: visual context tree is unchanged after clearing the highlight"
            : `FAIL: expected ${baseCount} nodes after clearing the highlight, got ${clearedCount}`);
    });
</script>
