Tests/LibWeb/Text/expected/inspector-overlay-visual-context-tree-growth.txt

PASS: recording with an inspector highlight does not modify the visual context tree
PASS: visual context tree is unchanged after clearing the highlight
